Move pipeline commands to a separate file
