+++ slackbot/bot_utils.py
@@ -1,10 +1,12 @@
 import string
 import requests
+import json5
+from collections import OrderedDict
 from secrets import token_hex
 import slack_sdk as slack
 from bot_info import slack_token, botid, workerid, broker_url
 from airflow_api import update_slack_connection, set_variable
-from kombu_helper import drain_messages
+from kombu_helper import drain_messages, peek_message
 
 
 def clear_queues():
@@ -76,3 +78,65 @@ def download_file(msg):
             return filetype, response.content.decode("ascii", "ignore")
         else:
             return None, None
+
+
+def download_json(msg):
+    filetype, content = download_file(msg)
+    if not content:
+        return None
+    if filetype == "Python":
+        scope = {}
+        try:
+            exec(content, scope)
+            if "submit_parameters" not in scope or not callable(scope["submit_parameters"]):
+                return None
+            payloads = scope['submit_parameters']()
+        except:
+            replyto(msg, "Cannot execute the `submit_parameters` function in the script")
+            replyto(msg, "{}".format(traceback.format_exc()))
+        #upload_param(msg, payloads)
+        return payloads
+    else: #if filetype == "JavaScript/JSON":
+        try:
+            json_obj = json5.loads(content, object_pairs_hook=OrderedDict)
+        except (ValueError, TypeError) as e:
+            replyto(msg, "Cannot load the json file: {}".format(str(e)))
+            return None
+        return json_obj
+
+
+def upload_param(msg, param):
+    sc = slack.WebClient(slack_token, timeout=300)
+    channel = msg['channel']
+    userid = msg['user']
+    thread_ts = msg['thread_ts'] if 'thread_ts' in msg else msg['ts']
+    sc.files_upload(
+        channels=channel,
+        filename="param.json",
+        filetype="javascript",
+        thread_ts=thread_ts,
+        content=json.dumps(param, indent=4),
+        initial_comment="<@{}> current parameters".format(userid)
+    )
+
+
+def guess_run_type(param):
+    if "WORKER_IMAGE" in param:
+        return "seg_run"
+    elif "CHUNKFLOW_IMAGE" in param:
+        return "inf_run"
+    else:
+        return None
+
+
+def latest_param_type():
+    json_obj = peek_message(broker_url, "seuronbot_payload")
+    if isinstance(json_obj, list):
+        param = json_obj[0]
+    else:
+        param = json_obj
+
+    if param:
+        return guess_run_type(param)
+    else:
+        return None
